Refactor: Convert course-admin/feedback route to TypeScript
